Added relationship between user and npcs in database and in rails.
Conducted test by adding user ID #1 to all existing npcs and then ran
user.npcs and npc.user in bin/rails console to test that the macros
were working.

***NEXT***

Going to add ownership syntax to the npc controller so this relationship
will be created when a user creates an npc.