Added ownership syntax to the NpcController.
I have it set to that an index or a show will work on all resources for
all users.

I have it set so that a create/post action will rely on the macro .build
method to establish ownership from now on as npc resources are created.

I have it so that the update and delete will only work if you are the
user that owned those resources. It does this by looking through only
the current user's npcs, not all npcs. So if you try to edit an npc that
you don't own, the request will get denied.