Successfully added user_name prop to user objects.
Did a bunch of testing associated with this and I was able to both add
after-the-fact and also add upon creation. You can see the prop in macro
relationships as well.

***NEXT***

I want to try to make it so you can log in with EITHER your username OR
your EMAIL.